incorporate Thomas feedback

+++ services/QuillLMS/db/migrate/20220705143703_add_minimum_and_maximum_grade_levels_to_activities.rb
@@ -2,7 +2,7 @@
 
 class AddMinimumAndMaximumGradeLevelsToActivities < ActiveRecord::Migration[5.2]
   def change
-    add_column :activities, :minimum_grade_level, :integer
-    add_column :activities, :maximum_grade_level, :integer
+    add_column :activities, :minimum_grade_level, :smallint
+    add_column :activities, :maximum_grade_level, :smallint
   end
 end
